Replace search terms with example queries on clicking
